# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mooda API",
        "description": "REST API Mooda untuk aplikasi mobile (Flutter). Semua endpoint terproteksi memakai Bearer token (Sanctum).",
        "contact": {
            "name": "Mooda Teknologi Indonesia",
            "email": "dev@mooda.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.mooda.id/api/v1",
            "description": "Produksi"
        },
        {
            "url": "http://127.0.0.1:8080/api/v1",
            "description": "Lokal (dev)"
        }
    ],
    "paths": {
...
        "/fnb/orders": {
            "get": {
                "tags": [
                    "F&B - Kasir"
                ],
                "summary": "Daftar pesanan (filter status/tanggal)",
                "operationId": "e6c3ea221955a1fa2c9dcebe425c1887",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "paid",
                                "void"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "F&B - Kasir"
                ],
                "summary": "Buat pesanan baru",
                "operationId": "1d8f22274b9d9bf0c21e90df904c2e85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "table_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "order_type": {
                                        "type": "string",
                                        "example": "dine_in",
                                        "enum": [
                                            "dine_in",
                                            "take_away"
                                        ]
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "menu_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "pedas"
                                                },
                                                "addon_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/fnb/orders/{id}": {
            "get": {
                "tags": [
                    "F&B - Kasir"
                ],
                "summary": "Detail pesanan (+ item & struk)",
                "operationId": "e0fde90fcc2d186a65505df2509eeb75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/fnb/orders/{id}/pay": {
            "post": {
                "tags": [
                    "F&B - Kasir"
                ],
                "summary": "Bayar pesanan (cash/QRIS)",
                "operationId": "e742561ec2541ae8cb0f047cde6daa62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "example": "cash",
                                        "enum": [
                                            "cash",
                                            "qris"
                                        ]
                                    },
                                    "paid": {
                                        "type": "integer",
                                        "example": 60000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Terbayar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/laundry/orders": {
            "get": {
                "tags": [
                    "Laundry - Kasir"
                ],
                "summary": "Daftar nota laundry",
                "operationId": "724cff170d5d4f6694072052da51c179",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Laundry - Kasir"
                ],
                "summary": "Buat nota laundry (harga server-side, VIP diskon)",
                "operationId": "2695e9a1133173c3f372e8632b523981",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "delivery": {
                                        "type": "string",
                                        "example": "pickup",
                                        "enum": [
                                            "pickup",
                                            "antar"
                                        ]
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "qty": {
                                                    "description": "kg/pcs/pasang sesuai satuan layanan",
                                                    "type": "number",
                                                    "example": 3.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/laundry/orders/{id}/advance": {
            "post": {
                "tags": [
                    "Laundry - Produksi"
                ],
                "summary": "Naikkan status ke tahap berikutnya",
                "operationId": "24ae8ff9d23df1e956c8a0256ab64c82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/shifts/current": {
            "get": {
                "tags": [
                    "Shift"
                ],
                "summary": "Shift aktif saat ini",
                "operationId": "bceb74d065e33de9e8cec12e0b8fd4ab",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Belum buka shift"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/shifts/open": {
            "post": {
                "tags": [
                    "Shift"
                ],
                "summary": "Buka shift",
                "operationId": "64d837fed3275680b0ab21def8a26416",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "opening_cash": {
                                        "type": "integer",
                                        "example": 200000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dibuka"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/shifts/close": {
            "post": {
                "tags": [
                    "Shift"
                ],
                "summary": "Tutup shift (rekap kas)",
                "operationId": "17f6881ab54a3a939d548fb1351e419b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "closing_cash": {
                                        "type": "integer",
                                        "example": 950000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ditutup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}